Added Xbox controller basic support

+++ core/src/com/gwel/entities/Planet.java
@@ -13,7 +13,7 @@ public class Planet {
 	public static final float MAX_RADIUS = 100.0f;
 	static final int MAX_SAT = 5;
 	
-	private int seed;
+	private long seed;
 	private Body body;
 	public Vector2 position;
 	private float mass;
@@ -26,7 +26,7 @@ public class Planet {
 	ArrayList<Satellite> satellites;
 	
 
-	public Planet(Vector2 pos, float rad, int seed) {
+	public Planet(Vector2 pos, float rad, long seed) {
 		position = pos.cpy();
 		radius = rad;
 		this.seed = seed;
